extracted mysqli flaky test
